requirements.txt
streamlit
transformers
torch
pdfplumber
nltk
gensim
scikit-learn
pypdf
google.generativeai
python-dotenv
